refactor(product-service): apply dash-case for files in mock/, rename files with array to the plural
